Bug Fixes: Replaced deprecated 'pcap_lookupdev' with 'pcap_findalldevs' in 'analize.c'. Resolved issues with repeated arithmetic operations and comparisons between different data types.

+++ protocols/application/dhcp.h
@@ -41,6 +41,6 @@
 #define DHCP_END_OPTIONS        0xff
 
 
-protocol_info dissect_dhcp(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_dhcp(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 #endif
