+++ protocols/network/icmp.h
@@ -23,7 +23,7 @@
 #define ICMP_LEN                8
 
 
-protocol_info dissect_icmp(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_icmp(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 /* use command "ping 8.8.8.8" to launch icmp packets in the local network. Than scan using netdump analize -filter "icmp" */
