+++ protocols/network/icmpv6.h
@@ -35,6 +35,6 @@
 #define ICMPV6_MESSAGE_REDIR_CODE   137
 
 
-protocol_info dissect_icmpv6(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_icmpv6(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 #endif
